requirements.txt
django
psycopg[binary]

djangorestframework
markdown

numpy
onnx
onnxruntime
pillow
